add storybook story and refactor hover color logic
